feat(sankey): move stories to the dedicated storybook workspace
