+++ storybook/package.json
@@ -31,6 +31,7 @@
     "@nivo/pie": "workspace:*",
     "@nivo/radar": "workspace:*",
     "@nivo/radial-bar": "workspace:*",
+    "@nivo/sankey": "workspace:*",
     "@nivo/tooltip": "workspace:*",
     "@nivo/waffle": "workspace:*",
     "@babel/preset-env": "^7.21.5",
@@ -52,6 +53,8 @@
     "@types/d3-scale": "^3.2.3",
     "styled-components": "^5.3.3",
     "d3-random": "^1.1.2",
-    "@types/d3-random": "^1.1.3"
+    "@types/d3-random": "^1.1.3",
+    "d3-sankey": "^0.12.3",
+    "@types/d3-sankey": "^0.11.2"
   }
 }
